fine_tune_service/requirements.txt
Flask
transformers
torch
pillow
torchvision
diffusers
tokenizers
scheduler
image_encoder
feature_extractor
unet
safety_checker
